move apify-properties.ts to utils/ dir
